+++ internal/auth0/connection/database/resource_test.go
@@ -1,4 +1,4 @@
-package connection_test
+package database_test
 
 import (
 	"fmt"
